refactor(scheduler): split scheduler.rs into a directory module
Convert the 1321-line `crates/runtime/src/scheduler.rs` into a
directory module:

  crates/runtime/src/scheduler/
  ├── mod.rs       (PeriodicTask trait, SchedulerCtx, spawn_scheduler,
  │                 run_scheduler_loop, in-file mod tests, constants)
  ├── tasks.rs     (run_due_tasks, dispatch_scheduler_turn_request,
  │                 publish_schedule_trigger, resolve_home_channel_tools,
  │                 compute_next_run)
  ├── heartbeat.rs (run_heartbeat — just orchestrates file read +
  │                 dispatch_scheduler_turn_request from tasks.rs)
  ├── prune.rs     (prune_conversation_events)
  └── reap.rs      (reap_stale_and_recover + REAP_STALE_INTERVAL +
                    STALE_CLAIM_TIMEOUT)

Production code is now distributed by concern, each file at a
reasonable size (16-320 lines). mod.rs's production code drops to
~220 lines (the loop machinery + trait + built-in task structs);
the remaining ~660 lines is the in-file `mod tests`, kept whole here
so the test mod's `use super::*;` still picks up everything via
re-exports at the top of mod.rs. A follow-up could split the tests
into a `scheduler/tests/` directory, but that's a separate PR.

cargo test -p assistant-runtime --lib reports 220 passing (same as
before); 18 scheduler tests pass unchanged. make lint clean.
